worker related refactoring code to allow other parts of the codebase to send events
