+++ lib/Echidna/Types/Campaign.hs
@@ -72,106 +72,6 @@ data CampaignConf = CampaignConf
     -- Only relevant if symExec is True
   }
 
-data WorkerType = FuzzWorker | SymbolicWorker deriving (Eq)
-
-type WorkerId = Int
-
-data CampaignEvent
-  = WorkerEvent WorkerId WorkerType WorkerEvent
-  | Failure String
-  | ReproducerSaved String -- filename
-
-data WorkerEvent
-  = TestFalsified !EchidnaTest
-  | TestOptimized !EchidnaTest
-  | NewCoverage { points :: !Int, numCodehashes :: !Int, corpusSize :: !Int, transactions :: [Tx] }
-  | SymNoNewCoverage
-  | SymVerified !String
-  | SymExecError !String
-  | TxSequenceReplayed FilePath !Int !Int
-  | TxSequenceReplayFailed FilePath Tx
-  | WorkerStopped WorkerStopReason
-  -- ^ This is a terminal event. Worker exits and won't push any events after
-  -- this one
-  deriving Show
-
-instance ToJSON WorkerEvent where
-  toJSON = \case
-    TestFalsified test -> toJSON test
-    TestOptimized test -> toJSON test
-    NewCoverage { points, numCodehashes, corpusSize } ->
-      object [ "coverage" .= points, "contracts" .= numCodehashes, "corpus_size" .= corpusSize]
-    SymNoNewCoverage -> object []
-    SymExecError err -> object [ "error" .= err ]
-    SymVerified name -> object [ "name" .= name ]
-    TxSequenceReplayed file current total ->
-      object [ "file" .= file, "current" .= current, "total" .= total ]
-    TxSequenceReplayFailed file tx ->
-      object [ "file" .= file, "tx" .= tx ]
-    WorkerStopped reason -> object [ "reason" .= show reason ]
-
-data WorkerStopReason
-  = TestLimitReached
-  | SymbolicExplorationDone
-  | SymbolicVerificationDone
-  | TimeLimitReached
-  | FastFailed
-  | Killed !String
-  | Crashed !String
-  deriving Show
-
-ppCampaignEvent :: CampaignEvent -> String
-ppCampaignEvent = \case
-  WorkerEvent _ _ e -> ppWorkerEvent e
-  Failure err -> err
-  ReproducerSaved f -> "Saved reproducer to " <> f
-
-ppWorkerEvent :: WorkerEvent -> String
-ppWorkerEvent = \case
-  TestFalsified test ->
-    "Test " <> T.unpack (showTest test) <> " falsified!"
-  TestOptimized test ->
-    let name = case test.testType of OptimizationTest n _ -> n; _ -> error "fixme"
-    in "New maximum value of " <> T.unpack name <> ": " <> show test.value
-  NewCoverage { points, numCodehashes, corpusSize } ->
-    "New coverage: " <> show points <> " instr, "
-      <> show numCodehashes <> " contracts, "
-      <> show corpusSize <> " seqs in corpus"
-  SymNoNewCoverage ->
-    "Symbolic execution finished with no new coverage."
-  SymVerified name ->
-    "Symbolic execution finished verifying contract " <> name <> " using a single symbolic transaction."
-  SymExecError err ->
-    "Symbolic execution failed because " <> err
-  TxSequenceReplayed file current total ->
-    "Sequence replayed from corpus file " <> file <> " (" <> show current <> "/" <> show total <> ")"
-  TxSequenceReplayFailed file tx ->
-    "WARNING: Sequence replay from corpus file " <> file <> " failed. " <>
-    "The destination contract is not deployed for this transaction: " <> show tx <> ". " <>
-    "Remove the file or the transaction to fix the issue."
-  WorkerStopped TestLimitReached ->
-    "Test limit reached. Stopping."
-  WorkerStopped SymbolicExplorationDone ->
-    "Symbolic worker ran out of transactions to explore. Stopping."
-  WorkerStopped SymbolicVerificationDone ->
-    "Symbolic worker finished with the list of methods to verify. Stopping."
-  WorkerStopped TimeLimitReached ->
-    "Time limit reached. Stopping."
-  WorkerStopped FastFailed ->
-    "A test was falsified. Stopping."
-  WorkerStopped (Killed e) ->
-    "Killed (" <> e <>"). Stopping."
-  WorkerStopped (Crashed e) ->
-    "Crashed:\n\n" <>
-    e <>
-    "\n\nPlease report it to https://github.com/crytic/echidna/issues"
-  where
-    showTest test = case test.testType of
-      PropertyTest n _ -> n
-      AssertionTest _ n _ -> encodeSig n
-      CallTest n _ -> n
-      _ -> error "impossible"
-
 -- | The state of a fuzzing campaign.
 data WorkerState = WorkerState
   { workerId    :: !Int
@@ -240,10 +140,3 @@ getNFuzzWorkers conf = maybe defaultN fromIntegral conf.workers
     n = numCapabilities
     maxN = max 1 n
     defaultN = min 4 maxN -- capped at 4 by default
-
--- | Number of workers, including SymExec worker if there is one
-getNWorkers :: CampaignConf -> Int
-getNWorkers conf = getNFuzzWorkers conf + (if conf.symExec then 1 else 0)
-
-workerIDToType :: CampaignConf -> WorkerId -> WorkerType
-workerIDToType conf wid = if conf.symExec && wid == (getNWorkers conf - 1) then SymbolicWorker else FuzzWorker
