+++ lib/Echidna/Types/Config.hs
@@ -14,7 +14,8 @@ import EVM.Types (Addr, Contract, W256)
 
 import Echidna.SourceAnalysis.Slither (SlitherInfo)
 import Echidna.SourceMapping (CodehashMap)
-import Echidna.Types.Campaign (CampaignConf, CampaignEvent)
+import Echidna.Types.Campaign (CampaignConf)
+import Echidna.Types.Worker (CampaignEvent)
 import Echidna.Types.Corpus (Corpus)
 import Echidna.Types.Coverage (CoverageMap)
 import Echidna.Types.Solidity (SolConf)
